Move branch from last snapshot using omni, correct location

+++ bibdesk/BDSKAddCommand.m
@@ -78,7 +78,7 @@
         id locationSpecifier = [[self arguments] objectForKey:@"ToLocation"];
         id insertionContainer = nil;
         NSString *insertionKey = nil;
-        NSInteger insertionIndex = -1;
+        int insertionIndex = -1;
         NSScriptClassDescription *containerClassDescription = nil;
         NSArray *classDescriptions = [insertionObjects valueForKey:@"scriptClassDescription"];
         NSScriptClassDescription *insertionClassDescription = [classDescriptions containsObject:[NSNull null]] ? nil : [NSScriptClassDescription commonAncestorForClassDescriptions:classDescriptions];
