+++ bibdesk/BDSKBorderlessKeyWindow.m
@@ -44,7 +44,7 @@
 @implementation BDSKBorderlessKeyWindow
 
 - (id)initWithContentRect:(NSRect)contentRect 
-                styleMask:(NSUInteger)aStyle
+                styleMask:(unsigned int)aStyle
                   backing:(NSBackingStoreType)bufferingType 
                     defer:(BOOL)flag {
