+++ bibdesk/BDSKFieldEditor.h
@@ -55,11 +55,11 @@
 @end
 
 // the above delegate methods could be implemented by calling these delegate methods for NSControl subclasses that actually have a delegate
-// currently implemented for NSTextField and NSTableView
+// currently implemented for NSTableView
 @interface NSControl (NSControlBDSKFieldEditorDelegate)
 - (NSRange)control:(NSControl *)control textView:(NSTextView *)textView rangeForUserCompletion:(NSRange)charRange;
 - (BOOL)control:(NSControl *)control textViewShouldAutoComplete:(NSTextView *)textView;
 - (BOOL)control:(NSControl *)control textViewShouldLinkKeys:(NSTextView *)textView;
 - (BOOL)control:(NSControl *)control textView:(NSTextView *)textView isValidKey:(NSString *)key;
-- (BOOL)control:(NSControl *)control textView:(NSTextView *)textView clickedOnLink:(id)aLink atIndex:(NSUInteger)charIndex;
+- (BOOL)control:(NSControl *)control textView:(NSTextView *)textView clickedOnLink:(id)aLink atIndex:(unsigned)charIndex;
 @end
